+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 4.43.0 (2025-04-01)
 
 - added: Sonic Mainnet network support
+- changed: (EVM) Allow minimum custom gas price to go below default if detected network gas price is lower
 
 ## 4.42.3 (2025-03-28)